Refactor: Extract PDF generation into generate_pdf.py module
- Add Medical Devices, Supply Deliveries, Psychosocial Scores, Insurance History, Clinical Notes sections
- Add dosing frequency, encounter duration, allergies field to PDF
